+++ src/pages/salary/SalaryAdmin.jsx
@@ -4,8 +4,8 @@ import { query, collection, where, getDocs } from "firebase/firestore";
 import { db } from "../../shared/firebase";
 import PageTitle from "../../shared/components/PageTitle";
 import SelectBox from "../../shared/components/SelectBox";
-import CalcBox from "../mySalary/components/CalcBox";
-import EditableCalcBox from "../mySalary/components/EditableCalcBox";
+import CalcBox from "../salary/components/CalcBox";
+import EditableCalcBox from "../salary/components/EditableCalcBox";
 import styled from "styled-components";
 import useAvailableMonths from "./hooks/useAvailableMonths";
 import useSalaryData from "./hooks/useSalaryData";
