+++ src/pages/salary/components/CalcBox.jsx
