+++ src/pages/salary/components/EditableCalcBox.jsx
