+++ src/shared/styles/GlobalStyle.jsx
